💄 style: move footer inside landing
