+++ src/app.css.ts
@@ -1,12 +1,15 @@
 import { style } from "@vanilla-extract/css";
 import { s } from "./styles";
+import * as colors from "./config/colors";
 
 export const container = style([
   s.flexColumn,
   s.flexSpaceBetween,
   s.scroll,
   {
-    minHeight: "100vh",
+    position: "fixed",
+    height: "100vh",
     width: "100vw",
+    backgroundColor: colors.PAPER,
   },
 ]);
