+++ src/landing/index.tsx
@@ -46,6 +46,9 @@ export function Landing() {
             Continue with GitHub
           </button>
         </div>
+        <footer className={css.footer}>
+          Every deployment from the 2023 edition of Jaam Toast.
+        </footer>
       </div>
     </div>
   );
